Create ConnectionBanner UI control and controller
The connection to the server is managed by a UI control that has the
controller that manages the connection to the server.

This allows for a unified point of contact with the server and also means
that the application can continues to run in the absense of the connection
(it just drops a banner to tell the user they are disconnected).