Create Login Event and fire when user enters login
This uses the event driven architecture, the controller fires the login
controller fires the event, it is filtered by the main controller which
sends it to the network and results would return to both the login
controller and main controller handlers.

Hopefully this design won't result in a weird Event spaghetti, but we will
find out at some point.