Fix AppScene reconnect bug
The client events sent to the server assume the server knows the current
user, if the user reconnects in the app scene the user information was
never being sent.

This commit fixes it so the app scene knows user info and logs in upon
reconnecting to tell the user the identity.